Challange-2/mofa's kitchen/requirements.txt
# requirements.txt
fastapi
uvicorn
pymongo
google-generativeai
python-dotenv
pydantic
motor
httpx
pytest
python-multipart
typing-extensions